+++ interface/src/GLCanvas.cpp
@@ -39,11 +39,11 @@ bool GLCanvas::isThrottleRendering() const {
 }
 
 int GLCanvas::getDeviceWidth() const {
-    return width() * (windowHandle() ? windowHandle()->devicePixelRatio() : 1.0f);
+    return width() * (windowHandle() ? (float)windowHandle()->devicePixelRatio() : 1.0f);
 }
 
 int GLCanvas::getDeviceHeight() const {
-    return height() * (windowHandle() ? windowHandle()->devicePixelRatio() : 1.0f);
+    return height() * (windowHandle() ? (float)windowHandle()->devicePixelRatio() : 1.0f);
 }
 
 void GLCanvas::initializeGL() {
